Add sac experiment for grasping.
[skip ci]

+++ magi/experimental/environments/bullet_kuka_env.py
@@ -126,7 +126,7 @@ class KukaDiverseObjectEnv(KukaGymEnv):
         self.action_space = spaces.Box(
             low=-1, high=1, shape=(4,))  # dx, dy, dz, da
     self.observation_space = spaces.Box(
-        low=0, high=255, shape=(self._height, self._width, 3))
+        low=0, high=255, shape=(self._height, self._width, 3), dtype=np.uint8)
     self.viewer = None
 
   def reset(self):
